+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 - Support for `Enum` attributes and `Content` properties
 - Improved performance of `.rbxm` parsing
+- `max_unsynced_changes` setting can now be set to much higher values
 
 ### Fixed
 
@@ -18,6 +19,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 ### Changed
 
 - Updated `rbx-dom` library to the latest major version
+- Config will now fail to load if `line_ending` setting is invalid
 
 ## [2.0.23] - 2025-02-05